Replace context by state for processor
A change in design. The context now goes inside the state. The
processors still receives an extra argument, but rather than that being
a context it's a more generally encompassing argument that currently has
both the context and the matches.

Some tests currently failing. They likely just need to be slightly changed
to reflect the new approach.