Start to test some basic schema scenarios
Added the `isint` and `isnotint` booleans. Both are active processors.
They lay the ground for the "onfail" and "cast" options. They'll also be
used as blueprint for other type checking processors.

Added flags to disable the native matching of booleans as integers and
vice-versa in the `is_intlike()` and `is_boolean()` utils functions.